requirements.txt
# External dependency for MySQL connectivity.
mysql-connector-python
